Move test to new file and add one more.
